Added a profile icon

+++ src/main/java/com/avelircraft/repositories/image/ImagesCrudRepository.java
@@ -4,6 +4,9 @@ import com.avelircraft.models.Image;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface ImagesCrudRepository extends CrudRepository<Image, Long> {
+public interface ImagesCrudRepository extends CrudRepository<Image, Integer> {
+
 }
